# Enable parallel execution of tests