--- a/cardano-cli/cardano-cli.cabal
+++ b/cardano-cli/cardano-cli.cabal
@@ -268,8 +268,11 @@ library cardano-cli-test-lib
                       , filepath
                       , hedgehog
                       , hedgehog-extras ^>= 0.6.1.0
+                      , lifted-base
+                      , monad-control
                       , process
                       , text
+                      , transformers-base
 
 test-suite cardano-cli-test
   import:               project-config
@@ -313,7 +316,7 @@ test-suite cardano-cli-test
                         Test.Cli.Pipes
                         Test.Cli.Shelley.Run.Query
 
-  ghc-options:          -threaded -rtsopts -with-rtsopts=-N -with-rtsopts=-T
+  ghc-options:          -threaded -rtsopts "-with-rtsopts=-N -T"
 
 test-suite cardano-cli-golden
   import:               project-config
@@ -420,4 +423,4 @@ test-suite cardano-cli-golden
                         Test.Golden.TxView
                         Test.Golden.Version
 
-  ghc-options:          -threaded -rtsopts -with-rtsopts=-N -with-rtsopts=-T
+  ghc-options:          -threaded -rtsopts "-with-rtsopts=-N -T"