# Changelog

```yaml
- description: |
    Enable parallel execution of tests
# uncomment types applicable to the change:
  type:
  # - feature        # introduces a new feature
  # - breaking       # the API has changed in a breaking way
  # - compatible     # the API has changed but is non-breaking
  # - optimisation   # measurable performance improvements
  # - improvement    # QoL changes e.g. refactoring
  # - bugfix         # fixes a defect
   - test           # fixes/modifies tests
  # - maintenance    # not directly related to the code
  # - release        # related to a new release preparation
  # - documentation  # change in code docs, haddocks...
```

# Context

See an explanation in similar PR in node: https://github.com/IntersectMBO/cardano-node/pull/5692

Before change:
```
All 204 tests passed (16.75s)
Test suite cardano-cli-golden: PASS
...
All 22 tests passed (3.94s)
Test suite cardano-cli-test: PASS
```
After change
```
All 204 tests passed (12.38s)
Test suite cardano-cli-golden: PASS
...
All 22 tests passed (1.97s)
Test suite cardano-cli-test: PASS
```

# Checklist

- [ ] Commit sequence broadly makes sense and commits have useful messages
- [ ] New tests are added if needed and existing tests are updated. See [Running tests](https://github.com/input-output-hk/cardano-node-wiki/wiki/Running-tests) for more details
- [ ] Self-reviewed the diff

<!-- 
### Note on CI ###
If your PR is from a fork, the necessary CI jobs won't trigger automatically for security reasons.
You will need to get someone with write privileges. Please contact IOG node developers to do this
for you. 
-->